Spec: Timeout contributions from Shared Storage when context ID is set (#102)
Prevents a Shared Storage operation from sending contributions after a
timeout, if a context ID was used. This will allow for reducing the
reporting delay for this case as it prevents the reporting time from
leaking any information about how long execution took (see issue #80 and
the corresponding explainer change #100).