Spec: Fix syntax error in "obtain the plaintext payload"
I was surprised to see errors on CI that appeared unrelated to my
changes in PR #171. I couldn't repro the errors until I upgraded from
bikeshed 4.2.8 to 5.0.0.

For future reference, the errors looked like this prior to fixing the
link syntax:

```
LINE 1126:8: Parsed a dfn autolink ([=...=]) inside of an <a>. Either close the <a> properly, or escape the autolink.
LINE 1127:24: Parsed a dfn autolink ([=...=]) inside of an <a>. Either close the <a> properly, or escape the autolink.
LINE 1124:27: [=...=] autolink opened at 1124:27 wasn't closed within 3 lines. You might have forgotten to close it; if not, switch to the HTML syntax to spread your link across that many lines.
LINE 1124: No 'dfn' refs found for 'filteringid}}"] and'.
[="{{PAHistogramContribution/filteringId}}"] and
=]
```